✨ Make ABI positions and sizes f32 for sub-pixel drawing

+++ examples/hello/src/lib.rs
@@ -5,31 +5,31 @@
 use rico8::*;
 
 struct Hello {
-    x: i32,
-    y: i32,
+    x: f32,
+    y: f32,
 }
 
 impl Game for Hello {
     fn update(&mut self, ctx: &mut Context) {
         if ctx.btn(Button::Left) {
-            self.x -= 1;
+            self.x -= 1.0;
         }
         if ctx.btn(Button::Right) {
-            self.x += 1;
+            self.x += 1.0;
         }
         if ctx.btn(Button::Up) {
-            self.y -= 1;
+            self.y -= 1.0;
         }
         if ctx.btn(Button::Down) {
-            self.y += 1;
+            self.y += 1.0;
         }
     }
 
     fn draw(&self, gfx: &mut Graphics) {
         gfx.clear(Color::BLACK);
-        gfx.print("hello, rico-8!", 36, 40, Color::WHITE);
-        gfx.rect_fill(self.x, self.y, 8, 8, Color::WHITE);
+        gfx.print("hello, rico-8!", 36.0, 40.0, Color::WHITE);
+        gfx.rect_fill(self.x, self.y, 8.0, 8.0, Color::WHITE);
     }
 }
 
-rico8::game!(Hello { x: 60, y: 64 });
+rico8::game!(Hello { x: 60.0, y: 64.0 });
