+++ examples/sprite_move/src/lib.rs
@@ -5,8 +5,8 @@
 use rico8::*;
 
 struct SpriteMove {
-    x: i32,
-    y: i32,
+    x: f32,
+    y: f32,
     flip: bool,
     walking: bool,
     frame: u32,
@@ -16,44 +16,44 @@ impl Game for SpriteMove {
     fn update(&mut self, ctx: &mut Context) {
         self.walking = false;
         if ctx.btn(Button::Left) {
-            self.x -= 1;
+            self.x -= 1.0;
             self.flip = true;
             self.walking = true;
         }
         if ctx.btn(Button::Right) {
-            self.x += 1;
+            self.x += 1.0;
             self.flip = false;
             self.walking = true;
         }
         if ctx.btn(Button::Up) {
-            self.y -= 1;
+            self.y -= 1.0;
             self.walking = true;
         }
         if ctx.btn(Button::Down) {
-            self.y += 1;
+            self.y += 1.0;
             self.walking = true;
         }
-        self.x = self.x.clamp(0, SCREEN_W - 8);
-        self.y = self.y.clamp(0, SCREEN_H - 8);
+        self.x = self.x.clamp(0.0, (SCREEN_W - 8) as f32);
+        self.y = self.y.clamp(0.0, (SCREEN_H - 8) as f32);
         self.frame += 1;
     }
 
     fn draw(&self, gfx: &mut Graphics) {
         gfx.clear(Color::DARK_BLUE);
-        gfx.print("arrows to walk", 36, 4, Color::LIGHT_GREY);
+        gfx.print("arrows to walk", 36.0, 4.0, Color::LIGHT_GREY);
         // Sprites 1 and 2 are the two walk frames.
         let frame = if self.walking && (self.frame / 4) % 2 == 0 {
             2
         } else {
             1
         };
-        gfx.spr_ext(SpriteId(frame), self.x, self.y, 1, 1, self.flip, false);
+        gfx.spr_ext(SpriteId(frame), self.x, self.y, 1.0, 1.0, self.flip, false);
     }
 }
 
 rico8::game!(SpriteMove {
-    x: 60,
-    y: 64,
+    x: 60.0,
+    y: 64.0,
     flip: false,
     walking: false,
     frame: 0,
